+++ src/Widgets/FullCalendarWidget.php
@@ -7,6 +7,7 @@ use Filament\Forms\Contracts\HasForms;
 use Filament\Widgets\Widget;
 use Illuminate\View\View;
 use Saade\FilamentFullCalendar\Widgets\Concerns\CanManageEvents;
+use Saade\FilamentFullCalendar\Widgets\Concerns\CanRefreshEvents;
 use Saade\FilamentFullCalendar\Widgets\Concerns\FiresEvents;
 use Saade\FilamentFullCalendar\Widgets\Concerns\UsesConfig;
 
@@ -16,6 +17,7 @@ class FullCalendarWidget extends Widget implements HasForms
         CanManageEvents::getForms insteadof InteractsWithForms;
     }
 
+    use CanRefreshEvents;
     use FiresEvents;
     use UsesConfig;