+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to `filament-fullcalendar` will be documented in this file.
 
+## v0.3.0 - 2022-06-27
+
+### What's Changed
+
+- Update FiresEvents.php by @wychoong in https://github.com/saade/filament-fullcalendar/pull/8
+- feat: allow opening or not events in a new tab by @saade in https://github.com/saade/filament-fullcalendar/pull/12
+- feat: refresh calendar events by @saade in https://github.com/saade/filament-fullcalendar/pull/13
+- fix: convert laravel locale to fullcalendar compatible locale by @saade in https://github.com/saade/filament-fullcalendar/pull/14
+
+### New Contributors
+
+- @wychoong made their first contribution in https://github.com/saade/filament-fullcalendar/pull/8
+
+**Full Changelog**: https://github.com/saade/filament-fullcalendar/compare/v0.2.1...v0.3.0
+
 ## v0.2.1 - 2022-02-21
 
 ## What's Changed